server/requirements.txt
# Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database — Unified MongoDB
beanie>=1.26.0
motor>=3.6.0
pymongo>=4.9.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI
google-generativeai>=0.8.0
openai>=1.50.0

# Utilities
pydantic[email]>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
cloudinary>=1.41.0
slowapi>=0.1.9
httpx>=0.27.0

# Email — verification + password reset + notifications
fastapi-mail>=1.4.0
